refactor(subs): use validator for continue operation as well
